+++ Functions/Assertions/PesterThrow.ps1
@@ -2,20 +2,20 @@
 # because this is a script block, the user will have to
 # wrap the code they want to assert on in { }
 function PesterThrow([scriptblock] $script, $expectedErrorMessage) {
-    $pester.ActualExceptionMessage = ""
-    $pester.ActualExceptionWasThrown = $false
+    $ActualExceptionMessage = ""
+    $ActualExceptionWasThrown = $false
 
     try {
         # Piping to Out-Null so results of script exeution
         # does not remain on the pipeline
         & $script | Out-Null
     } catch {
-        $pester.ActualExceptionWasThrown = $true
-        $pester.ActualExceptionMessage = $_.Exception.Message
+        $ActualExceptionWasThrown = $true
+        $ActualExceptionMessage = $_.Exception.Message
     }
 
-    if ($pester.ActualExceptionWasThrown) {
-        return (Get-DoMessagesMatch $pester.ActualExceptionMessage $expectedErrorMessage)
+    if ($ActualExceptionWasThrown) {
+        return (Get-DoMessagesMatch $ActualExceptionMessage $expectedErrorMessage)
     }
     return $false
 }
@@ -28,7 +28,7 @@ function Get-DoMessagesMatch($value, $expected) {
 function PesterThrowFailureMessage($value, $expected) {
     if ($expected) {
       return "Expected: the expression to throw an exception with message {{{0}}}, an exception was {2}raised, message was {{{1}}}" -f
-              $expected, $pester.ActualExceptionMessage,(@{$true="";$false="not "}[$pester.ActualExceptionWasThrown])
+              $expected, $ActualExceptionMessage,(@{$true="";$false="not "}[$ActualExceptionWasThrown])
     } else {
       return "Expected: the expression to throw an exception"
     }
@@ -37,9 +37,9 @@ function PesterThrowFailureMessage($value, $expected) {
 function NotPesterThrowFailureMessage($value, $expected) {
     if ($expected) {
         return "Expected: the expression to not throw an exception with message {{{0}}}, an exception was {2}raised, message was {{{1}}}" -f
-              $expected, $pester.ActualExceptionMessage,(@{$true="";$false="not "}[$pester.ActualExceptionWasThrown])
+              $expected, $ActualExceptionMessage,(@{$true="";$false="not "}[$ActualExceptionWasThrown])
     } else {
-        return "Expected: the expression to not throw an exception. Message was {{{0}}}" -f $pester.ActualExceptionMessage
+        return "Expected: the expression to not throw an exception. Message was {{{0}}}" -f $ActualExceptionMessage
     }
 }