+++ Functions/TestResults.Tests.ps1
@@ -5,90 +5,68 @@ Describe "Write nunit test results" {
     Setup -Dir "Results"
 
     It "should write a successful test result" {
-        $testResults = @{}
-        $testResults.Describes = @{
-            name = 'Mocked Describe'
-            Tests = @{
-                testNumber = 1
-                name = "Successful testcase"
-                time = "1.0"
-                success = $true
-            };
-        }
-
+				#create state
+				$TestResults = New-PesterState -Path TestDrive:\
+				$testResults.EnterDescribe('Mocked Describe')
+				$TestResults.AddTestResult("Successful testcase",$true,(New-TimeSpan -Seconds 1))
+				
+				#export and validate the file
         $testFile = "$TestDrive\Results\Tests.xml"
-        Write-NunitTestReport $testResults $testFile
+        Export-NunitReport $testResults $testFile
         $xmlResult = [xml] (Get-Content $testFile)
         $xmlTestCase = $xmlResult.'test-results'.'test-suite'.'results'.'test-case'
         $xmlTestCase.name     | Should Be "Successful testcase"
         $xmlTestCase.result   | Should Be "Success"
-        $xmlTestCase.time     | Should Be "1.0"
+        $xmlTestCase.time     | Should Be "1"
     }
 
     It "should write a failed test result" {
-        $testResults = @{}
-        $testResults.Describes = @{
-            name = 'Mocked Describe'
-            Tests = @{
-                name = "Failed testcase"
-                time = "2.0"
-                failureMessage = 'Assert failed: "Expected: Test. But was: Testing"';
-                stackTrace = 'at line: 28 in  C:\Pester\Result.Tests.ps1'
-                success = $false
-            };
-        }
-
+				#create state
+				$TestResults = New-PesterState -Path TestDrive:\
+				$testResults.EnterDescribe('Mocked Describe')
+				$time = [TimeSpan]25000000 #2.5 seconds
+				$TestResults.AddTestResult("Failed testcase",$false,$time,'Assert failed: "Expected: Test. But was: Testing"','at line: 28 in  C:\Pester\Result.Tests.ps1')
+				
+				#export and validate the file
         $testFile = "$TestDrive\Results\Tests.xml"
-        Write-NunitTestReport $testResults $testFile
+        Export-NunitReport $testResults $testFile
         $xmlResult = [xml] (Get-Content $testFile)
         $xmlTestCase = $xmlResult.'test-results'.'test-suite'.'results'.'test-case'
         $xmlTestCase.name                   | Should Be "Failed testcase"
         $xmlTestCase.result                 | Should Be "Failure"
-        $xmlTestCase.time                   | Should Be "2.0"
+        $xmlTestCase.time                   | Should Be "2.5"
         $xmlTestCase.failure.message        | Should Be 'Assert failed: "Expected: Test. But was: Testing"'
         $xmlTestCase.failure.'stack-trace'  | Should Be 'at line: 28 in  C:\Pester\Result.Tests.ps1'
 
     }
 
      It "should write the test summary" {
-        $testResults = @{}
-        $testResults.Describes = @{
-            name = 'Mocked Describe'
-            Tests = @{
-                name = "Testcase"
-                time = "1.0"
-                success = $true
-            };
-        }
+				#create state
+				$TestResults = New-PesterState -Path TestDrive:\
+				$testResults.EnterDescribe('Mocked Describe')
+				$TestResults.AddTestResult("Testcase",$true,(New-TimeSpan -Seconds 1))
+				
+				#export and validate the file
         $testFile = "$TestDrive\Results\Tests.xml"
-        Write-NunitTestReport $testResults $testFile
+        Export-NunitReport $testResults $testFile
         $xmlResult = [xml] (Get-Content $testFile)
         $xmlTestResult = $xmlResult.'test-results'
-        $xmlTestResult.total    | Should Be 0
+        $xmlTestResult.total    | Should Be 1
         $xmlTestResult.failures | Should Be 0
         $xmlTestResult.date     | Should Be $true
         $xmlTestResult.time     | Should Be $true
     }
 
     it "should write the test-suite information" {
-        $testResults = @{}
-        $testResults.Describes = @{
-                name = 'Mocked Describe'
-                Tests = @(
-                    @{
-                        name = "Successful testcase"
-                        time = "1.0"
-                        success = $true
-                    },
-                    @{
-                        name = "Successful testcase"
-                        time = "1.1"
-                        success = $true
-                    };
-                );
-        }
+				#create state
+				$TestResults = New-PesterState -Path TestDrive:\
+				$testResults.EnterDescribe('Mocked Describe')
+				$TestResults.AddTestResult("Successful testcase",$true,[timespan]10000000) #1.0 seconds
+				$TestResults.AddTestResult("Successful testcase",$true,[timespan]11000000) #1.1 seconds
+				
+				#export and validate the file
         $testFile = "$TestDrive\Results\Tests.xml"
-        Write-NunitTestReport $testResults $testFile
+        Export-NunitReport $testResults $testFile
         $xmlResult = [xml] (Get-Content $testFile)
 
         $xmlTestResult = $xmlResult.'test-results'.'test-suite'
@@ -100,28 +78,17 @@ Describe "Write nunit test results" {
     }
 
     it "should write two test-suite elements for two describes" {
-        $testResults = @{}
-        $testResults.Describes = @( 
-            @{
-                name = 'Describe #1'
-                Tests =  @{
-                    name = "Successful testcase"
-                    time = "1.0"
-                    success = $true
-                }
-            },
-            @{
-                name = 'Describe #2'
-                Tests = @{
-                    name = "Failed testcase"
-                    time = "2.0"
-                    success = $false
-                }
-            }
-        );
-
+				#create state
+				$TestResults = New-PesterState -Path TestDrive:\
+				$testResults.EnterDescribe('Describe #1')
+				$TestResults.AddTestResult("Successful testcase",$true,(New-TimeSpan -Seconds 1))
+				$TestResults.LeaveDescribe()
+				$testResults.EnterDescribe('Describe #2')
+				$TestResults.AddTestResult("Failed testcase",$false,(New-TimeSpan -Seconds 2))
+				
+				#export and validate the file
         $testFile = "$TestDrive\Results\Tests.xml"
-        Write-NunitTestReport $testResults $testFile
+        Export-NunitReport $testResults $testFile
         $xmlResult = [xml] (Get-Content $testFile)
 
         $xmlTestSuite1 = $xmlResult.'test-results'.'test-suite'[0]
@@ -140,7 +107,7 @@ Describe "Write nunit test results" {
         $testResults = @{}
         $testResults.Tests = @( "" );
         $testFile = "$TestDrive\Results\Tests.xml"
-        Write-NunitTestReport $testResults $testFile
+        Export-NunitReport $testResults $testFile
         $xmlResult = [xml] (Get-Content $testFile)
 
         $xmlEnvironment = $xmlResult.'test-results'.'environment'
@@ -154,7 +121,7 @@ Describe "Write nunit test results" {
     }
 
     it "Should validate test results against the nunit 2.5 schema" {
-        $testResults = @{}
+				$testResults = @{}
         $testResults.Describes = @( 
             @{
                 name = 'Describe #1'
@@ -179,7 +146,7 @@ Describe "Write nunit test results" {
             }
         );
         $testFile = "$TestDrive\Results\Tests.xml"
-        Write-NunitTestReport $testResults $testFile
+        Export-NunitReport $testResults $testFile
         [xml]$xml = gc $testFile 
         # This has been failing for a while! It was previous checking a boolean
         # but it was actually a list of validation errors
@@ -216,10 +183,18 @@ Describe "Get-TestTime" {
 		#on cs-CZ, de-DE and other systems where decimal separator is ",". value [double]3.5 is output as 3,5 
 		#this makes some of the tests fail, it could also leak to the nUnit report if the time was output
 				
-		$TestResult = New-Object -TypeName psObject -Property @{ Time = 3.5 }
-		
+		$TestResult = New-Object -TypeName psObject -Property @{ Time = [timespan]35000000 } #3.5 seconds
+ 		
 		#using the string formatter here to know how the string will be output to screen
 		$Result = { Get-TestTime -Tests $TestResult | Out-String -Stream } | Using-Culture -Culture de-DE
 		$Result | Should Be "3.5"
+	}
+	It "Time is measured in seconds with 0,1 millisecond as lowest value" {
+			$TestResult = New-Object -TypeName psObject -Property @{ Time = [timespan]1000 } 
+			Get-TestTime -Tests $TestResult | Should Be 0.0001
+			$TestResult = New-Object -TypeName psObject -Property @{ Time = [timespan]100 } 
+			Get-TestTime -Tests $TestResult | Should Be 0
+			$TestResult = New-Object -TypeName psObject -Property @{ Time = [timespan]1234567 } 
+			Get-TestTime -Tests $TestResult | Should Be 0.1235
 	}
 }