+++ Functions/TestResults.ps1
@@ -1,109 +1,107 @@
-$scriptRoot = Split-Path $MyInvocation.MyCommand.Path -Parent
-function Get-GlobalTestResults {
-    if ($pester.testResults -ne $null) {
-        return $pester.testResults
-    }
-
-    $pester.testResults = @{}
-    $pester.testResults.Describes = @();
-    $pester.testResults.CurrentDescribe = @{ name = ''; Tests = @() }
-    $pester.testResults.TestCount = 0
-    $pester.testResults.FailedTestsCount = 0
-    $pester.testResults.TestDepth = 0
-    $pester.testResults.TotalTime = 0;
-
-    return $pester.testResults
-}
-
-function Reset-GlobalTestResults {
-    $pester.testResults = $null
-}
-
-function Write-TestReport {
-    $results = $pester.testResults
-    Write-Host "Tests completed in $(Get-HumanTime $results.TotalTime)"
-    Write-Host "Passed: $($results.TestCount - $results.FailedTestsCount) Failed: $($results.FailedTestsCount)"
-}
-
-function Get-HumanTime($seconds) {
-    if($seconds -gt 0.99) {
-        $time = [math]::Round($seconds, 2)
+function Get-HumanTime($Seconds) {
+    if($Seconds -gt 0.99) {
+        $time = [math]::Round($Seconds, 2)
         $unit = "s"
     }
     else {
-        $time = [math]::Floor($seconds * 1000)
+        $time = [math]::Floor($Seconds * 1000)
         $unit = "ms"
     }
     return "$time$unit"
 }
 
-function Write-NunitTestReport($results, $outputFile) {
-    if($results -eq $null) {
-        return;
-    }
-    $report = @{
-        runDate = (Get-Date -format "yyyy-MM-dd")
-        runTime = (Get-Date -format "HH:mm:ss")
-        total = 0;
-        failures = 0;
-    }
-
-    $report.total = $results.TestCount
-    $report.failures = $results.FailedTestsCount
-    $report.TestSuites = (Get-TestSuites $results.Describes)
-    $report.testCases = (Get-TestResults $results.Tests)
-    $report.Environment = (Get-RunTimeEnvironment)
-    Invoke-Template 'TestResults.template.xml' $report | Set-Content $outputFile -force
+function ConvertTo-NUnitReport {
+  param (
+    [parameter(Mandatory=$true,ValueFromPipeline=$true)]
+    [PSObject[]]$InputObject
+  )
+	#TODO clean this mess up
+	$results = $InputObject
+	
+	$outputFile = $Path
+  $report = @{
+    runDate = (Get-Date -format "yyyy-MM-dd")
+    runTime = (Get-Date -format "HH:mm:ss")
+    total = 0;
+    failures = 0;
+	}
+	$report.total = $results.TotalCount
+  $report.failures = $results.FailedCount
+  $report.TestSuites = (Get-TestSuites $results.TestResult)
+  $report.Environment = (Get-RunTimeEnvironment)
+	$report = New-Object -TypeName Psobject -Property $report
+	Invoke-Template 'TestResults.template.xml' $report 
+}
+function Export-NUnitReport {
+  param (
+    [parameter(Mandatory=$true,ValueFromPipeline=$true)]
+    [PSObject[]]$InputObject,
+		[parameter(Mandatory=$true)]
+    [String]$Path
+	)
+	
+	ConvertTo-NUnitReport -InputObject $InputObject | Set-Content $Path -Force
 }
 
-function Get-TestSuites($describes) {
-    $testSuites = ( $describes | %{
+function Get-TestSuites ($describes) {
+    $describes | Group-Object -Property Describe | foreach {
         $suite = @{  
             resultMessage = "Failure"
             totalTime = "0.0"
             name = $_.name
         }
-        $suite.testCases = (Get-TestResults $_.Tests)  
-        $suite.totalTime = (Get-TestTime $_.Tests)
-        $suite.success = (Get-TestSuccess $_.Tests)
+				#calculate the time first, I am converting the time into string in the TestCases
+				$suite.totalTime = (Get-TestTime $_.Group)
+        $suite.testCases = (Get-TestResults $_.Group)  
+        $suite.success = (Get-TestSuccess $_.Group)
         if($suite.success -eq "True") 
         {
             $suite.resultMessage = "Success" 
         }
         Invoke-Template 'TestSuite.template.xml' $suite
-    })
+    }
     return $testSuites
 }
 
+function Convert-TimeSpan {
+	param (
+		[Parameter(Mandatory=$true,ValueFromPipeline=$true)]
+		[TimeSpan]$TimeSpan
+	)
+	process {
+		[string][math]::round($TimeSpan.totalseconds,4)
+	}
+}
 function Get-TestTime($tests) {
-    $totalTime = 0;
+    [TimeSpan]$totalTime = 0;
     $tests | %{
         $totalTime += $_.time
     }
-    return [string]$totalTime;
+    $totalTime | Convert-TimeSpan
 }
 
 function Get-TestSuccess($tests) {
-    $success = "True"
-    $tests | %{
-        if($_.success -eq $false) {
-            $success = "False"
-        }
+    #if any fails, the whole suite fails
+		$result = $true
+		$tests | foreach {
+			if (-not $_.Passed) {
+				$result = $false
+			}
     }
-    return $success;
+		[String]$result
 }
 
-function Get-TestResults($tests) {
-    $testCaseXmls = $tests | %{ 
-        $result = $_
-        if($result.success) {
-            Invoke-Template 'TestCaseSuccess.template.xml' $result
-        }
-        else {
-            Invoke-Template  'TestCaseFailure.template.xml' $result
-        }
-    }
-    return $testCaseXmls
+function Get-TestResults($Test) {
+	$Test | %{ 
+	  $result = $_
+		$result.Time = $result.Time | Convert-TimeSpan
+	  if($result.Passed) {
+	    Invoke-Template 'TestCaseSuccess.template.xml' $result
+	  }
+	  else {
+	    Invoke-Template  'TestCaseFailure.template.xml' $result
+	  }
+	}
 }
 
 function Get-RunTimeEnvironment() {
@@ -123,8 +121,7 @@ function Get-RunTimeEnvironment() {
 
 
 
-function Exit-WithCode {
-    $failedTestsCount = $Pester.TestResults.FailedTestsCount
-    $host.SetShouldExit($failedTestsCount)
+function Exit-WithCode ($FailedCount) {
+    $host.SetShouldExit($FailedCount)
 }